10-10CG | Clean up app constants
